chore: merge branch 'master' into Jia/1744-Refactor-the-images-folder-structure
