+++ src/components/custom/other-platforms.tsx
@@ -22,7 +22,7 @@ import { DbotIcon } from 'images/svg/dbot'
 import DMT5 from 'images/svg/dmt5/dmt5-icon.svg'
 import BinaryBot from 'images/svg/binarybot-icon.svg'
 import DTrader from 'images/svg/dtrader/dtrader-icon.svg'
-import DerivGo from 'images/svg/deriv-go/deriv-go-icon.svg'
+import DerivGo from 'images/svg/components/component_deriv_go_icon.svg'
 import DerivLife from 'images/svg/menu/deriv-life.svg'
 import Forex from 'images/svg/components/component_forex_nav.svg'
 import Help from 'images/svg/menu/help-center.svg'
