+++ src/components/elements/off-canvas-menu.tsx
@@ -33,7 +33,7 @@ import DMT5 from 'images/svg/dmt5/dmt5-icon.svg'
 import DerivX from 'images/svg/components/component_deriv_x.svg'
 import { DerivEzLogoRed } from 'images/svg/deriv-ez'
 import DTrader from 'images/svg/dtrader/dtrader-icon.svg'
-import DerivGo from 'images/svg/deriv-go/deriv-go-icon.svg'
+import DerivGo from 'images/svg/components/component_deriv_go_icon.svg'
 import Forex from 'images/svg/components/component_forex_nav.svg'
 import Help from 'images/svg/menu/help-center.svg'
 import CFD from 'images/svg/components/component_margin_trading_nav.svg'
