+++ src/features/components/templates/navigation/payment-agent-nav/index.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react'
 import NavTemplate from '../template'
 import affiliateNavItems from './data'
 import { partners_nav_logo, partners_buttons } from './payment-agent-nav.module.scss'
-import PrimeLogo from 'images/svg/deriv-prime/deriv-prime-logo.svg'
+import PrimeLogo from 'images/svg/deriv-prime/deriv_prime_logo.svg'
 import PartnerNavLogo from 'images/svg/partner-nav-logo.svg'
 import Image from 'features/components/atoms/image'
 import Button from 'features/components/atoms/button'
