+++ src/images/svg/dbot/index.ts
@@ -1,10 +1,10 @@
-export { default as DbotIcon } from './dbot-icon.svg'
-export { default as BotIcon } from './bot-icon.svg'
-export { default as DbotBgMobile } from './dbot-bg-mobile.svg'
-export { default as DbotBg } from './dbot-bg.svg'
-export { default as BotBannerLogo } from './deriv-bot-banner-logo.svg'
+export { default as DbotIcon } from './dbot_icon.svg'
+export { default as BotIcon } from './dbot_bot_icon.svg'
+export { default as DbotBgMobile } from './dbot_bg_mobile.svg'
+export { default as DbotBg } from './dbot_bg.svg'
+export { default as BotBannerLogo } from './dbot_banner_logo.svg'
 export { default as DtIcon } from '../dtrader_icon.svg'
 export { default as DxIcon } from '../deriv_x_icon.svg'
-export { default as EzIcon } from './ez-icon.svg'
+export { default as EzIcon } from './dbot_ez_icon.svg'
 export { default as GoIcon } from '../deriv_go_icon.svg'
 export { default as Mt5Icon } from './mt5.svg'
