+++ src/images/svg/deriv-ez/index.ts
@@ -1,11 +1,11 @@
-export { default as DbIcon } from './db-icon.svg'
-export { default as DerivEzLogoRed } from './deriv-ez-logo-red.svg'
-export { default as DerivEzLogo } from './deriv-ez-logo.svg'
-export { default as DerivEzQR } from './derivez-qr.svg'
-export { default as Dmt5Icon } from './dmt5-icon.svg'
+export { default as DbIcon } from './deriv_ez_db_icon.svg'
+export { default as DerivEzLogoRed } from './deriv_ez_logo_red.svg'
+export { default as DerivEzLogo } from './deriv_ez_logo.svg'
+export { default as DerivEzQR } from './deriv_ez_qr.svg'
+export { default as Dmt5Icon } from './deriv_ez_dmt5_icon.svg'
 export { default as DtIcon } from '../dtrader_icon.svg'
 export { default as DxIcon } from '../deriv_x_icon.svg'
-export { default as EasyToStart } from './easy-to-start.svg'
-export { default as FavouriteAssets } from './favourite-assets.svg'
+export { default as EasyToStart } from './deriv_ez_easy_to_start.svg'
+export { default as FavouriteAssets } from './deriv_ez_favourite_assets.svg'
 export { default as DGoIcon } from '../deriv_go_icon.svg'
-export { default as TwentyFourSeven } from './twenty-four-seven.svg'
+export { default as TwentyFourSeven } from './deriv_ez_twenty_four_seven.svg'
