+++ src/images/svg/deriv-prime/index.ts
@@ -1,9 +1,9 @@
-import BrokerageFirms from './brokerage_firms.svg'
-import InvestmentFunds from './investment_funds.svg'
-import AssetManagers from './asset_managers.svg'
-import ProfessionalTraders from './professional_traders.svg'
-import FinancialInstitutions from './financial_institutions.svg'
-import CorporateClients from './corporate_clients.svg'
+import BrokerageFirms from './deriv_prime_brokerage_firms.svg'
+import InvestmentFunds from './deriv_prime_investment_funds.svg'
+import AssetManagers from './deriv_prime_asset_managers.svg'
+import ProfessionalTraders from './deriv_prime_professional_traders.svg'
+import FinancialInstitutions from './deriv_prime_financial_institutions.svg'
+import CorporateClients from './deriv_prime_corporate_clients.svg'
 
 export {
     BrokerageFirms,
