+++ src/pages/check-email/_icon-grid.tsx
@@ -4,10 +4,10 @@ import device from 'themes/device'
 import { Text } from 'components/elements'
 import { Localize, localize } from 'components/localization'
 // Icons
-import Spam from 'images/svg/check-email/spam.svg'
-import Typo from 'images/svg/check-email/typo.svg'
-import WorkEmail from 'images/svg/check-email/work-email.svg'
-import Firewalls from 'images/svg/check-email/firewalls.svg'
+import Spam from 'images/svg/check-email/check_email_spam.svg'
+import Typo from 'images/svg/check-email/check_email_typo.svg'
+import WorkEmail from 'images/svg/check-email/check_email_work_email.svg'
+import Firewalls from 'images/svg/check-email/check_email_firewalls.svg'
 import { Container } from 'components/containers'
 import { TString } from 'types/generics'
