+++ src/pages/deriv-go/_banner.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { StaticImage } from 'gatsby-plugin-image'
 import CommonHeaderSection from 'components/elements/common-header-section'
-import DerivGOLogo from 'images/svg/deriv-go/deriv-go-banner-logo.svg'
+import DerivGOLogo from 'images/svg/deriv-go/deriv_go_banner_logo.svg'
 import device from 'themes/device'
 import useBreakpoints from 'components/hooks/use-breakpoints'
 import useHandleSignup from 'components/hooks/use-handle-signup'
