+++ src/pages/deriv-go/_footer-banner.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import derivGoLogo from '../../images/svg/deriv-go/derivGo.svg'
-import derivGoQR from '../../images/svg/deriv-go/derivGoQR.svg'
+import derivGoLogo from '../../images/svg/deriv-go/deriv_go.svg'
+import derivGoQR from '../../images/svg/deriv-go/deriv_go_qr.svg'
 import GrayAngle30 from '../../images/common/gray-angle.png'
 import AppleIcon from '../../images/svg/apple-icon.svg'
 import AppGalleryIcon from '../../images/svg/appGallery-icon.svg'
