+++ src/pages/deriv-go/_other-apps.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
-import mt5Icon from '../../images/svg/deriv-go/mt5.svg'
-import dtIcon from '../../images/svg/deriv-go/dt-icosn.svg'
-import ezIcon from '../../images/svg/deriv-go/ez-icon.svg'
+import mt5Icon from '../../images/svg/deriv-go/deriv_go_mt5.svg'
+import dtIcon from '../../images/svg/deriv-go/deriv_go_dt_icon.svg'
+import ezIcon from '../../images/svg/deriv-go/deriv_go_ez_icon.svg'
 import dxIcon from '../../images/svg/deriv_x_icon.svg'
-import dbIcon from '../../images/svg/deriv-go/db-icon.svg'
+import dbIcon from '../../images/svg/deriv-go/deriv_go_db_icon.svg'
 import GenericCarousel, { TRenderableData } from 'components/custom/carousel/_platforms-carousel'
 
 const carouselData: TRenderableData[] = [
